pyproject.toml
[tool.black]
line-length = 79
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "data_pipeline_practice"
version = "0.1.0"
description = "Data Pipeline Sandbox"
authors = ["Blake Enyart <blake.enyart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.aws-s3" = "^1.83.0"
"aws-cdk.aws-apigatewayv2-integrations" = "^1.83.0"
"aws-cdk.aws-apigatewayv2" = "^1.83.0"
"aws-cdk.aws-apigateway" = "^1.83.0"
"aws-cdk.aws-ecs-patterns" = "^1.83.0"
"aws-cdk.aws-ecs" = "^1.83.0"
requests = "^2.25.0"
boto3 = "^1.16.25"
"aws-solutions-constructs.core" = "^1.83.0"
"aws-solutions-constructs.aws-kinesis-streams-kinesis-firehose-s3" = "^1.83.0"
"aws-cdk.aws-logs" = "1.83.0"
"aws-cdk.aws-glue" = "1.83.0"
"aws-cdk.aws-kinesis" = "1.83.0"
"aws-cdk.pipelines" = "1.83.0"
"aws-cdk.aws-ecr" = "^1.83.0"
"aws-cdk.core" = "1.83.0"
"aws-cdk.aws-kinesisfirehose" = "1.83.0"
pytest = "^6.1.2"
klaxon = "^1.1.1"
invoke = "^1.5.0"
pre-commit = "^2.10.0"
black = "^20.8b1"
"aws-solutions-constructs.aws-dynamodb-stream-lambda" = "^1.83.0"
"aws-cdk.aws-lambda" = "1.83.0"
chalice = "^1.22.0"
"aws-cdk.aws-kms" = "1.83.0"
websockets = "^8.1"


[tool.poetry.dev-dependencies]
pandas = "^1.1.4"
SQLAlchemy = "^1.3.20"
ddlgenerator = "^0.1.9"
botocore = "^1.19.25"
fastparquet = "^0.4.1"
pylint = "^2.6.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
